// ui/src/utils/dashboard_client/api.ts
export interface ClusterinfoAlertManagerInfo {
    /**
     * 
     * @type {string}
     * @memberof ClusterinfoAlertManagerInfo
     */
    binary_path?: string;
    /**
     * 
     * @type {string}
     * @memberof ClusterinfoAlertManagerInfo
     */
    ip?: string;
    /**
     * 
     * @type {number}
     * @memberof ClusterinfoAlertManagerInfo
     */
    port?: number;
}
/**
 * 
 * @export
 * @interface ClusterinfoClusterInfo
 */
export interface ClusterinfoClusterInfo {
    /**
     * 
     * @type {ClusterinfoAlertManagerInfo}
     * @memberof ClusterinfoClusterInfo
     */
    alert_manager?: ClusterinfoAlertManagerInfo;
    /**
     * 
     * @type {ClusterinfoGrafanaInfo}
     * @memberof ClusterinfoClusterInfo
     */
    grafana?: ClusterinfoGrafanaInfo;
    /**
     * 
     * @type {ClusterinfoClusterInfoPd}
     * @memberof ClusterinfoClusterInfo
     */
    pd?: ClusterinfoClusterInfoPd;
    /**
     * 
     * @type {ClusterinfoClusterInfoTidb}
     * @memberof ClusterinfoClusterInfo
     */
    tidb?: ClusterinfoClusterInfoTidb;
    /**
     * 
     * @type {ClusterinfoClusterInfoTikv}
     * @memberof ClusterinfoClusterInfo
     */
    tikv?: ClusterinfoClusterInfoTikv;
}
/**
 * 
 * @export
 * @interface ClusterinfoClusterInfoPd
 */
export interface ClusterinfoClusterInfoPd {
    /**
     * 
     * @type {string}
     * @memberof ClusterinfoClusterInfoPd
     */
    err?: string;
    /**
     * 
     * @type {Array<ClusterinfoPDInfo>}
     * @memberof ClusterinfoClusterInfoPd
     */
    nodes?: Array<ClusterinfoPDInfo>;
}
/**
 * 
 * @export
 * @interface ClusterinfoClusterInfoTidb
 */
export interface ClusterinfoClusterInfoTidb {
    /**
     * 
     * @type {string}
     * @memberof ClusterinfoClusterInfoTidb
     */
    err?: string;
    /**
     * 
     * @type {Array<ClusterinfoTiDBInfo>}
     * @memberof ClusterinfoClusterInfoTidb
     */
    nodes?: Array<ClusterinfoTiDBInfo>;
}
/**
 * 
 * @export
 * @interface ClusterinfoClusterInfoTikv
 */
export interface ClusterinfoClusterInfoTikv {
    /**
     * 
     * @type {string}
     * @memberof ClusterinfoClusterInfoTikv
     */
    err?: string;
    /**
     * 
     * @type {Array<ClusterinfoTiKVInfo>}
     * @memberof ClusterinfoClusterInfoTikv
     */
    nodes?: Array<ClusterinfoTiKVInfo>;
}
/**
 * 
 * @export
 * @interface ClusterinfoGrafanaInfo
 */
export interface ClusterinfoGrafanaInfo {
    /**
     * 
     * @type {string}
     * @memberof ClusterinfoGrafanaInfo
     */
    binary_path?: string;
    /**
     * 
     * @type {string}
     * @memberof ClusterinfoGrafanaInfo
     */
    ip?: string;
    /**
     * 
     * @type {number}
     * @memberof ClusterinfoGrafanaInfo
     */
    port?: number;
}
/**
 * 
 * @export
 * @interface ClusterinfoPDInfo
 */
export interface ClusterinfoPDInfo {
    /**
     * 
     * @type {string}
     * @memberof ClusterinfoPDInfo
     */
    deploy_path?: string;
    /**
     * 
     * @type {string}
     * @memberof ClusterinfoPDInfo
     */
    ip?: string;
    /**
     * 
     * @type {number}
     * @memberof ClusterinfoPDInfo
     */
    port?: number;
    /**
     * 
     * @type {number}
     * @memberof ClusterinfoPDInfo
     */
    status?: number;
    /**
     * 
     * @type {string}
     * @memberof ClusterinfoPDInfo
     */
    version?: string;
}
/**
 * 
 * @export
 * @interface ClusterinfoTiDBInfo
 */
export interface ClusterinfoTiDBInfo {
    /**
     * 
     * @type {string}
     * @memberof ClusterinfoTiDBInfo
     */
    binary_path?: string;
    /**
     * 
     * @type {string}
     * @memberof ClusterinfoTiDBInfo
     */
    ip?: string;
    /**
     * 
     * @type {number}
     * @memberof ClusterinfoTiDBInfo
     */
    port?: number;
    /**
     * 
     * @type {number}
     * @memberof ClusterinfoTiDBInfo
     */
    status?: number;
    /**
     * 
     * @type {number}
     * @memberof ClusterinfoTiDBInfo
     */
    status_port?: number;
    /**
     * 
     * @type {string}
     * @memberof ClusterinfoTiDBInfo
     */
    version?: string;
}
/**
 * 
 * @export
 * @interface ClusterinfoTiKVInfo
 */
export interface ClusterinfoTiKVInfo {
    /**
     * 
     * @type {string}
     * @memberof ClusterinfoTiKVInfo
     */
    binary_path?: string;
    /**
     * 
     * @type {string}
     * @memberof ClusterinfoTiKVInfo
     */
    ip?: string;
    /**
     * 
     * @type {{ [key: string]: string; }}
     * @memberof ClusterinfoTiKVInfo
     */
    labels?: { [key: string]: string; };
    /**
     * 
     * @type {number}
     * @memberof ClusterinfoTiKVInfo
     */
    port?: number;
    /**
     * 
     * @type {number}
     * @memberof ClusterinfoTiKVInfo
     */
    status?: number;
    /**
     * 
     * @type {number}
     * @memberof ClusterinfoTiKVInfo
     */
    status_port?: number;
    /**
     * 
     * @type {string}
     * @memberof ClusterinfoTiKVInfo
     */
    version?: string;
}
// ...
export interface DiagnosticspbSearchLogRequest {
    /**
     * 
     * @type {number}
     * @memberof DiagnosticspbSearchLogRequest
     */
    end_time?: number;
    /**
     * 
     * @type {Array<number>}
     * @memberof DiagnosticspbSearchLogRequest
     */
    levels?: Array<number>;
    /**
     * We use a string array to represent multiple CNF pattern sceniaor like: SELECT * FROM t WHERE c LIKE \'%s%\' and c REGEXP \'.*a.*\' because Golang and Rust don\'t support perl-like (?=re1)(?=re2)
     * @type {Array<string>}
     * @memberof DiagnosticspbSearchLogRequest
     */
    patterns?: Array<string>;
    /**
     * 
     * @type {number}
     * @memberof DiagnosticspbSearchLogRequest
     */
    start_time?: number;
}
// ...
export interface LogsearchSearchTarget {
    /**
     * 
     * @type {string}
     * @memberof LogsearchSearchTarget
     */
    ip?: string;
    /**
     * 
     * @type {number}
     * @memberof LogsearchSearchTarget
     */
    kind?: number;
    /**
     * 
     * @type {number}
     * @memberof LogsearchSearchTarget
     */
    port?: number;
    /**
     * 
     * @type {number}
     * @memberof LogsearchSearchTarget
     */
    status_port?: number;
}
/**
 * 
 * @export
 * @interface LogsearchTaskGroupModel
 */
export interface LogsearchTaskGroupModel {
    /**
     * 
     * @type {number}
     * @memberof LogsearchTaskGroupModel
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof LogsearchTaskGroupModel
     */
    log_store_dir?: string;
    /**
     * 
     * @type {DiagnosticspbSearchLogRequest}
     * @memberof LogsearchTaskGroupModel
     */
    search_request?: DiagnosticspbSearchLogRequest;
    /**
     * 
     * @type {number}
     * @memberof LogsearchTaskGroupModel
     */
    state?: number;
}
/**
 * 
 * @export
 * @interface LogsearchTaskGroupResponse
 */
export interface LogsearchTaskGroupResponse {
    /**
     * 
     * @type {LogsearchTaskGroupModel}
     * @memberof LogsearchTaskGroupResponse
     */
    task_group?: LogsearchTaskGroupModel;
    /**
     * 
     * @type {Array<LogsearchTaskModel>}
     * @memberof LogsearchTaskGroupResponse
     */
    tasks?: Array<LogsearchTaskModel>;
}
/**
 * 
 * @export
 * @interface LogsearchTaskModel
 */
export interface LogsearchTaskModel {
    /**
     * 
     * @type {string}
     * @memberof LogsearchTaskModel
     */
    error?: string;
    /**
     * 
     * @type {number}
     * @memberof LogsearchTaskModel
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof LogsearchTaskModel
     */
    log_store_path?: string;
    /**
     * 
     * @type {LogsearchSearchTarget}
     * @memberof LogsearchTaskModel
     */
    search_target?: LogsearchSearchTarget;
    /**
     * 
     * @type {number}
     * @memberof LogsearchTaskModel
     */
    state?: number;
    /**
     * 
     * @type {number}
     * @memberof LogsearchTaskModel
     */
    task_group_id?: number;
}